+++ src/osd/modules/sound/sdl_sound.c
@@ -184,7 +184,7 @@ void sound_sdl::unlock_buffer(void)
 
 void sound_sdl::att_memcpy(void *dest, const INT16 *data, int bytes_to_copy)
 {
-	int level= (int) (pow(10.0, (float) attenuation / 20.0) * 128.0);
+	int level= (int) (pow(10.0, (double) attenuation / 20.0) * 128.0);
 	INT16 *d = (INT16 *) dest;
 	int count = bytes_to_copy/2;
 	while (count>0)
